requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pyyaml>=6.0
jupyter>=1.0
ipykernel>=6.0
control>=0.9
pytest>=7.0
